Fix bug

See merge request propensi-fasilkom-ui-2023/kelas-a/a05/a05-be!54

+++ src/main/java/com/a05/simaya/usaha/service/UsahaServiceImpl.java
@@ -74,7 +74,9 @@ public class UsahaServiceImpl implements UsahaService {
         if (usaha != null){
             List<GambarUsahaModel> gambarUsahaModels = gambarUsahaDb.findAllByUsahaModel(usaha);
             CatatanModel catatanModel = catatanDb.findCatatanModelByUsaha(usaha);
-            catatanDb.delete(catatanModel);
+            if(catatanModel != null){
+                catatanDb.delete(catatanModel);
+            }
             gambarUsahaDb.deleteAll(gambarUsahaModels);
             usahaDb.delete(usahaModel);
             return true;
